requirements.txt
alembic==1.4.2
attrs==19.3.0
autopep8==1.5.3
Babel==2.8.0
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coverage==5.2.1
dominate==2.5.1
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
Flask-HTTPAuth==4.1.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
httpie==2.2.0
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
more-itertools==8.4.0
mysql-connector-python==8.0.21
numpy==1.19.1
opencv-python==4.3.0.36
packaging==20.4
pluggy==0.13.1
protobuf==3.12.2
py==1.9.0
pycodestyle==2.6.0
pycrypto==2.6.1
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.0.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
urllib3==1.25.10
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
